+++ gauss.py
@@ -184,12 +184,6 @@ def matrixmoltiplication(matA:list[list[int]],matB:list[list[int]])->list[list[i
         matout.append(lout)
     return matout
 
-def markov(vet:list[Fraction],mat:list[list[Fraction]], n:int) -> list[Fraction]:
-    v = [vet]
-    print(v)
-    for i in range(n):
-        v = matrixmoltiplication(v,mat)
-    return v
 
 # classica matrice quadrata
 square_matrix = [[1.5,3,1,-1],[3,9,4,1],[2,1,5,2],[0,1,-1,-1]]
@@ -204,9 +198,3 @@ invmatrix1 = [[1,2],[2,3]]
 matrix2 = [[1,2,-1],[-2,0,1],[1,-1,0]]
 invmatrix2 = [[1,1,2],[1,1,1],[2,3,4]]
 
-# matrici per markov
-
-markovmat =to_fractions([['1/2','1/4','1/4'],['1/4','1/2','1/4'],['1/4','1/4','1/2']])
-markovvet = [1,0,0]
-
-printMat(markov(markovvet,markovmat,3))
